refactor: extract shared settings metadata builders from nvhttp
The builders that describe settings state to clients lived in nvhttp's
anonymous namespace, which meant the web UI's HTTP server could not
reuse them and reimplemented the same logic in JavaScript. This moves
the mode-selection facades, the stream display mode options JSON, the
auto-quality policy builder and its blocked-reason mapping, the host
virtual-display availability check, and a new build_tuning_json
(extracted verbatim from the session-status handler's inline block)
into a shared settings_metadata unit. Bodies are moved unchanged and
nvhttp's call sites are qualified; nvhttp's client-visible JSON stays
byte-identical except one additive key.

That key is badge: mode_option_t now carries the stream_path registry
descriptor's badge, so the host's real badge vocabulary, including
Recommended on headless_stream, reaches the mode options JSON instead
of being dropped at the projection boundary. The non-Linux fallback
branch mirrors the registry values.

Two source-guard tests pinned the moved bodies by their nvhttp source
text; they are re-pointed at settings_metadata.cpp with their
invariants intact, including the guard that keeps the two-argument
effective-selection overload from reading live proc session state.
New tests cover the blocked-reason matrix, the off-state policy
contract that is the only reachable state while the legacy AI flag is
hardcoded off, the tuning JSON field set, and badge propagation.
